HIP-584: Refactoring estimate gas tests: removing hardcoded values (#6017)
Rather than relying on fixed expected values for the estimate, the approach involves initiating a regular transaction simulation using MirrorEvmTxProcessorFacade. The resulting gasUsed value from this simulation is obtained and serves as a reference for estimation. Subsequently, the eth_estimateGas request is then checked whether its offset is between 5% and 20% bigger than the gasUsed value.

In order to be more accurate with the estimation the MirrorNodeEvmProperties.estimateGasIterationThreshold values is changed from 1200 to 3600 .

Signed-off-by: Yanko Nikolov <nikolovqnko@gmail.com>
Signed-off-by: Steven Sheehy <steven.sheehy@swirldslabs.com>
Co-authored-by: Steven Sheehy <steven.sheehy@swirldslabs.com>