HIP-584: Contract create estimations are treated as call estimations (#6056)
Estimate gas for contract creation is treated as contract call estimate. This is caused by a wrong gas calculator configuration.

To resolve it a simple flag for indicating contract creations was changed in MirrorEvmTxProcessor. Knowing that address zero is set as default receiver in estimate contract create cases, we use it as a condition check.

Signed-off-by: Yanko Nikolov <nikolovqnko@gmail.com>
Signed-off-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>
Co-authored-by: IvanKavaldzhiev <ivankavaldzhiev@gmail.com>